FIX: 특정 상황에서 프로젝트가 삭제되지 않는 문제

+++ src/main/java/bssm/deploy/domain/project/service/ProjectService.java
@@ -111,8 +111,14 @@ public class ProjectService {
         User user = currentUser.getCachedUser();
         Project project = projectProvider.findByIdAndUser(projectId, user);
 
-        cancelDeployService.cancelDeployProject(CancelDeployProjectReq.ofProjectId(projectId));
-        containerService.removeContainer(project);
+        try {
+            cancelDeployService.cancelDeployProject(CancelDeployProjectReq.ofProjectId(projectId));
+        } catch (Exception ignored) {}
+        try {
+            if (project.checkContainerProject()) {
+                containerService.removeContainer(project);
+            }
+        } catch (Exception ignored) {}
         processProjectFileService.deleteProjectFile(project);
         projectRepository.delete(project);
     }
